Modify Ui diagram and refactor Ui packages
