+++ src/main/java/team/serenity/ui/LessonDataPanel.java
@@ -1,4 +1,4 @@
-package team.serenity.ui.lessondata;
+package team.serenity.ui;
 
 import java.util.logging.Logger;
 
@@ -12,9 +12,7 @@ import team.serenity.commons.core.LogsCenter;
 import team.serenity.commons.core.Messages;
 import team.serenity.model.group.question.Question;
 import team.serenity.model.group.studentinfo.StudentInfo;
-import team.serenity.ui.DataPanel;
-import team.serenity.ui.groupdata.GroupDataPanel;
-import team.serenity.ui.serenitydata.SerenityDataPanel.QuestionListViewCell;
+import team.serenity.ui.SerenityDataPanel.QuestionListViewCell;
 
 public class LessonDataPanel extends DataPanel {
     private static final String FXML = "LessonDataPanel.fxml";
