+++ src/main/java/team/serenity/ui/StudentInfoCard.java
@@ -1,4 +1,4 @@
-package team.serenity.ui.lessondata;
+package team.serenity.ui;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
@@ -7,7 +7,6 @@ import javafx.scene.image.ImageView;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
 import team.serenity.model.group.studentinfo.StudentInfo;
-import team.serenity.ui.UiPart;
 
 /**
  * An UI component that displays information of a {@code StudentInfo}.
